+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Onboarding flow UI updates
+- Update WC to check for optional namespaces when approving session
 
 ## [Released 3.3.11]
 - fix PLT value formatter